Issue #121: Initial work on dynamic fields.
